# perf: conditionally import Telemetry