@ijjk 2024-03-21
## Failing test suites <!--21978586699-->
Commit: d1a572810e117ec43e41092c8a5c708aa481e518

`pnpm test test/integration/app-dir-export/test/dynamicpage-prod.test.ts`
- app dir - with output export - dynamic api route prod > production mode > should work in prod with dynamicPage undefined
- app dir - with output export - dynamic api route prod > production mode > should work in prod with dynamicPage 'error'
- app dir - with output export - dynamic api route prod > production mode > should work in prod with dynamicPage 'force-static'
- app dir - with output export - dynamic api route prod > production mode > should work in prod with dynamicPage 'force-dynamic'

<details>
<summary>Expand output</summary>

  ● app dir - with output export - dynamic api route prod › production mode › should work in prod with dynamicPage undefined

    thrown: "Exceeded timeout of 60000 ms for a test.
    Add a timeout value to this test to increase the timeout, if this is a long-running test. See https://jestjs.io/docs/api#testname-fn-timeout."

      12 |           'Page with `dynamic = "force-dynamic"` couldn\'t be exported. `output: "export"` requires all pages be renderable statically',
      13 |       },
    > 14 |     ])(
         |      ^
      15 |       'should work in prod with dynamicPage $dynamicPage',
      16 |       async ({ dynamicPage, expectedErrMsg }) => {
      17 |         await runTests({ isDev: false, dynamicPage, expectedErrMsg })

      at ../node_modules/.pnpm/jest-each@29.7.0/node_modules/jest-each/build/bind.js:47:15
          at Array.forEach (<anonymous>)
      at integration/app-dir-export/test/dynamicpage-prod.test.ts:14:6
      at integration/app-dir-export/test/dynamicpage-prod.test.ts:4:52
      at Object.describe (integration/app-dir-export/test/dynamicpage-prod.test.ts:3:1)

  ● app dir - with output export - dynamic api route prod › production mode › should work in prod with dynamicPage 'error'

    Failed to replace content.

    Pattern: const dynamic = 'force-static'

    Content: import Link from 'next/link'

    export const dynamic = undefined

    export function generateStaticParams() {
      return [{ slug: 'first' }, { slug: 'second' }]
    }

    export default function Page({ params }) {
      return (
        <main>
          <h1>{params.slug}</h1>
          <ul>
            <li>
              <Link href="/another">Visit another page</Link>
            </li>
          </ul>
        </main>
      )
    }

      669 |     } else if (typeof pattern === 'string') {
      670 |       if (!currentContent.includes(pattern)) {
    > 671 |         throw new Error(
          |               ^
      672 |           `Failed to replace content.\n\nPattern: ${pattern}\n\nContent: ${currentContent}`
      673 |         )
      674 |       }

      at File.replace (lib/next-test-utils.ts:671:15)
      at replace (integration/app-dir-export/test/utils.ts:112:14)
      at integration/app-dir-export/test/dynamicpage-prod.test.ts:17:23

  ● app dir - with output export - dynamic api route prod › production mode › should work in prod with dynamicPage 'force-static'

    Failed to replace content.

    Pattern: const dynamic = 'force-static'

    Content: import Link from 'next/link'

    export const dynamic = undefined

    export function generateStaticParams() {
      return [{ slug: 'first' }, { slug: 'second' }]
    }

    export default function Page({ params }) {
      return (
        <main>
          <h1>{params.slug}</h1>
          <ul>
            <li>
              <Link href="/another">Visit another page</Link>
            </li>
          </ul>
        </main>
      )
    }

      669 |     } else if (typeof pattern === 'string') {
      670 |       if (!currentContent.includes(pattern)) {
    > 671 |         throw new Error(
          |               ^
      672 |           `Failed to replace content.\n\nPattern: ${pattern}\n\nContent: ${currentContent}`
      673 |         )
      674 |       }

      at File.replace (lib/next-test-utils.ts:671:15)
      at replace (integration/app-dir-export/test/utils.ts:112:14)
      at integration/app-dir-export/test/dynamicpage-prod.test.ts:17:23

  ● app dir - with output export - dynamic api route prod › production mode › should work in prod with dynamicPage 'force-dynamic'

    Failed to replace content.

    Pattern: const dynamic = 'force-static'

    Content: import Link from 'next/link'

    export const dynamic = undefined

    export function generateStaticParams() {
      return [{ slug: 'first' }, { slug: 'second' }]
    }

    export default function Page({ params }) {
      return (
        <main>
          <h1>{params.slug}</h1>
          <ul>
            <li>
              <Link href="/another">Visit another page</Link>
            </li>
          </ul>
        </main>
      )
    }

      669 |     } else if (typeof pattern === 'string') {
      670 |       if (!currentContent.includes(pattern)) {
    > 671 |         throw new Error(
          |               ^
      672 |           `Failed to replace content.\n\nPattern: ${pattern}\n\nContent: ${currentContent}`
      673 |         )
      674 |       }

      at File.replace (lib/next-test-utils.ts:671:15)
      at replace (integration/app-dir-export/test/utils.ts:112:14)
      at integration/app-dir-export/test/dynamicpage-prod.test.ts:17:23

</details>


Read more about building and testing Next.js in [contributing.md](https://github.com/vercel/next.js/blob/canary/contributing.md).